refactor(cli): extract createBridgedEmitter and standardize no-active-runbook
Extract duplicated createBridgedEmitter helper to shared module and
standardize no-active-runbook error handling across all commands.

Changes:
- Create helpers/execution-emitter.ts with shared createBridgedEmitter
- Update pass, fail, goto, run commands to import shared helper
- Standardize no-active-runbook handling using output.noActiveRunbook()
- Pass action parameter to noActiveRunbook() for JSON output compatibility
- Fix goto.ts error code: UNKNOWN_ERROR -> ENGINE_INIT_FAILED
