+++ packages/cli/src/commands/goto.ts
@@ -8,38 +8,13 @@ import {
   parseStepIdFromString,
   stepIdToString,
   countNumberedSteps,
-  ExecutionEventEmitter,
-  type RunbookState,
 } from '@rundown-org/core';
 import { resolveRunbookFile } from '../helpers/resolve-runbook.js';
 import { getCwd } from '../helpers/context.js';
 import { runExecutionLoop } from '../services/execution.js';
 import { withErrorHandling } from '../helpers/wrapper.js';
 import { OutputEmitter } from '../services/output-emitter.js';
-
-/**
- * Create an event emitter for a runbook execution and bridge to OutputEmitter.
- *
- * @param runbookState - The runbook state to create the emitter for
- * @param output - The OutputEmitter to bridge events to
- * @returns The ExecutionEventEmitter
- */
-function createBridgedEmitter(
-  runbookState: RunbookState,
-  output: OutputEmitter
-): ExecutionEventEmitter {
-  const emitter = new ExecutionEventEmitter(
-    runbookState.id,
-    { name: runbookState.runbook, path: runbookState.runbookPath }
-  );
-
-  // Bridge execution events to the unified output system
-  emitter.subscribe((event) => {
-    output.executionEvent(event);
-  });
-
-  return emitter;
-}
+import { createBridgedEmitter } from '../helpers/execution-emitter.js';
 
 /**
  * Registers the 'goto' command for jumping to specific steps.
@@ -58,7 +33,7 @@ export function registerGotoCommand(program: Command): void {
         const state = await manager.getActive();
 
         if (!state) {
-          output.status(false, 'goto', 'No active runbook');
+          output.noActiveRunbook('goto');
           output.flush();
           return;
         }
@@ -136,7 +111,7 @@ export function registerGotoCommand(program: Command): void {
         // Create XState actor
         const actor = await manager.createActor(state.id, steps);
         if (!actor) {
-          output.error('Failed to initialize runbook engine', 'UNKNOWN_ERROR');
+          output.error('Failed to initialize runbook engine', 'ENGINE_INIT_FAILED');
           output.flush();
           process.exit(1);
         }
