+++ packages/cli/src/commands/run.ts
@@ -13,36 +13,13 @@ import {
   getErrorMessage,
   type PendingStep,
   type RunbookState,
-  ExecutionEventEmitter,
+  type ExecutionEventEmitter,
 } from '@rundown-org/core';
 import { resolveRunbookFile } from '../helpers/resolve-runbook.js';
 import { getCwd } from '../helpers/context.js';
 import { runExecutionLoop } from '../services/execution.js';
 import { OutputEmitter } from '../services/output-emitter.js';
-
-/**
- * Create an event emitter for a runbook execution and bridge to OutputEmitter.
- *
- * @param runbookState - The runbook state to create the emitter for
- * @param output - The OutputEmitter to bridge events to
- * @returns The ExecutionEventEmitter
- */
-function createBridgedEmitter(
-  runbookState: RunbookState,
-  output: OutputEmitter
-): ExecutionEventEmitter {
-  const emitter = new ExecutionEventEmitter(
-    runbookState.id,
-    { name: runbookState.runbook, path: runbookState.runbookPath }
-  );
-
-  // Bridge execution events to the unified output system
-  emitter.subscribe((event) => {
-    output.executionEvent(event);
-  });
-
-  return emitter;
-}
+import { createBridgedEmitter } from '../helpers/execution-emitter.js';
 
 /**
  * Emit RUNBOOK_STARTED event with metadata.
